fbfg

+++ supervised_learning/attention/4-positional_encoding.py
@@ -2,7 +2,6 @@
 '''cfvdfdfbfgbgbg'''
 
 import numpy as np
-import tensorflow as tf
 
 
 def positional_encoding(max_seq_len, dm):
@@ -17,4 +16,4 @@ def positional_encoding(max_seq_len, dm):
 
     pos_encoding = angle_rads[...]
 
-    return tf.cast(pos_encoding, dtype=tf.float32)
+    return pos_encoding
